add validations to project.sh and create basic docs

+++ project.sh
@@ -23,26 +23,38 @@ branch_push() {
 
 feature_start() {
 	if [[ $1 != '' ]]; then
-		git flow feature start $1
+		if [[ $BRANCH_CURRENT == $BRANCH_DEVELOP ]]; then
+			git flow feature start $1
+		else
+			echo "You should be at $BRANCH_DEVELOP branch to start new features."
+		fi
 	else
 		echo "You should enter feature name"
 	fi
 }
 
 feature_send() {
-	branch_push
-	gh pr create \
-		--head $BRANCH_CURRENT \
-		--base $BRANCH_DEVELOP \
-		--title "Merge $BRANCH_CURRENT to $BRANCH_DEVELOP" \
-		--web
+	if [[ $BRANCH_CURRENT == feature* ]]; then
+		branch_push
+		gh pr create \
+			--head $BRANCH_CURRENT \
+			--base $BRANCH_DEVELOP \
+			--title "Merge $BRANCH_CURRENT to $BRANCH_DEVELOP" \
+			--web
+	else
+		echo "You should be at a feature branch to send changes."
+	fi
 }
 
 feature_done() {
-	git checkout $BRANCH_DEVELOP
-	git branch -D $BRANCH_CURRENT
-	git push $GIT_REMOTE_NAME --delete $BRANCH_CURRENT
-	branch_pull $BRANCH_DEVELOP
+	if [[ $BRANCH_CURRENT == feature* ]]; then
+		git checkout $BRANCH_DEVELOP
+		git branch -D $BRANCH_CURRENT
+		git push $GIT_REMOTE_NAME --delete $BRANCH_CURRENT
+		branch_pull $BRANCH_DEVELOP
+	else
+		echo "You should be at a feature branch to done it."
+	fi
 }
 
 hotfix_start() {
@@ -54,19 +66,27 @@ hotfix_start() {
 }
 
 hotfix_send() {
-	branch_push
-	gh pr create \
-		--head $BRANCH_CURRENT \
-		--base $BRANCH_MASTER \
-		--title "Merge $BRANCH_CURRENT to $BRANCH_MASTER" \
-		--web
+	if [[ $BRANCH_CURRENT == hotfix* ]]; then
+		branch_push
+		gh pr create \
+			--head $BRANCH_CURRENT \
+			--base $BRANCH_MASTER \
+			--title "Merge $BRANCH_CURRENT to $BRANCH_MASTER" \
+			--web
+	else
+		echo "You should be at a hotfix branch to send changes."
+	fi
 }
 
 hotfix_done() {
-	git checkout $BRANCH_MASTER
-	git branch -D $BRANCH_CURRENT
-	git push $GIT_REMOTE_NAME --delete $BRANCH_CURRENT
-	branch_pull $BRANCH_MASTER
+	if [[ $BRANCH_CURRENT == hotfix* ]]; then
+		git checkout $BRANCH_MASTER
+		git branch -D $BRANCH_CURRENT
+		git push $GIT_REMOTE_NAME --delete $BRANCH_CURRENT
+		branch_pull $BRANCH_MASTER
+	else
+		echo "You should be at a hotfix branch to done it."
+	fi
 }
 
 # main
